Beta: Split into 2 separate kits, Democracy and Multisig
