+++ kits/democracy/.gitignore
@@ -1,0 +1,2 @@
+abis.json
+index_local.js
