+++ kits/democracy/contracts/Imports.sol
@@ -1,0 +1,10 @@
+/* solium-disable */
+pragma solidity 0.4.24;
+
+// HACK to workaround truffle artifact loading on dependencies
+
+import "@aragon/os/contracts/factory/ENSFactory.sol";
+import "@aragon/os/contracts/factory/APMRegistryFactory.sol";
+import "@aragon/os/contracts/factory/EVMScriptRegistryFactory.sol";
+
+contract Imports {}
