+++ kits/multisig/.soliumignore
@@ -1,0 +1,1 @@
+
