+++ kits/multisig/contracts/test/mocks/ExecutionTarget.sol
@@ -1,0 +1,22 @@
+pragma solidity 0.4.24;
+
+
+contract ExecutionTarget {
+    uint public counter;
+
+    function execute() public {
+        counter += 1;
+        emit Executed(counter);
+    }
+
+    function failExecute() public pure {
+        // solium-disable-next-line error-reason
+        revert();
+    }
+
+    function setCounter(uint x) public {
+        counter = x;
+    }
+
+    event Executed(uint x);
+}
