+++ kits/multisig/scripts/extract-abis.js
@@ -1,5 +1,4 @@
 const abis = {
-  DemocracyTemplate: require('../build/contracts/DemocracyTemplate').abi,
   MultisigTemplate: require('../build/contracts/MultisigTemplate').abi,
 }
